Aca separe el codigo
Hay algunos bugs cuando se cargan filtros y categorias, pero por lo menos el codigo esta mas ordenado. Con categorias y filtros simples anda bien, si se complica un poco ya capaz que tira un error
